requirements.txt
# Required packages to run TaIGa
pandas
biopython
